pyproject.toml
[tool.poetry]
name = "reddit-connectivity"
version = "0.1.0"
description = ""
authors = ["cjber <cjberragan@gmail.com>"]
readme = "README.md"
packages = [{ include = "reddit_connectivity" }]

[tool.poetry.dependencies]
python = "3.10.9"
pytorch-lightning = "^1.9.2"
datasets = "^2.9.0"
spacy = "^3.5.0"
seqeval = "^1.2.2"
psaw = "^0.1.0"
rtree = "^1.0.1"
scikit-learn = "^1.2.1"
networkx = "^3.0"
seaborn = "^0.12.2"
jsonlines = "^3.1.0"
numba = "^0.56.4"
h3 = "^3.7.6"
h3pandas = "^0.2.3"
nltk = "^3.8.1"
statsmodels = "^0.13.5"
polars = "^0.16.7"
scienceplots = "^2.0.1"
mapclassify = "^2.5.0"
wheel = "^0.38.4"
ipython = "^8.10.0"
pandas-stubs = "^1.5.3.230214"
pytest = "^7.2.1"
ipdb = "^0.13.11"
pyqt5 = "^5.15.9"
tabulate = "^0.9.0"
doccano = "^1.8.3"
dvc = "^2.45.1"
mlflow = "^2.1.1"
dvc-http = { git = "https://github.com/iterative/dvc-http.git" }
ruff = "^0.0.247"
neovim = "^0.3.1"
transformers = "^4.26.1"
torch = "^1.13.1"
jupyter = "^1.0.0"
dvc-s3 = "^2.21.0"
dagshub = "^0.2.12"
gradio = "^3.19.1"
plotly = "^5.13.0"
ruff-lsp = "^0.0.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
